requirements.txt
requests
bs4
colorama
requests_html
rich
python-dotenv
